Split logger functionality into two classes
