+++ src/index.ts
@@ -1,5 +1,7 @@
+import ExceptionLogger from './ExceptionLogger';
 import Logger from './Logger';
 
 export {
+  ExceptionLogger,
   Logger,
 };
